Update Table.java
Customers can not longer stay more than 4 hours in the restaurant, thanks to the new system that counts and increases the time a table is occupied, to free it automatically when the time increases.
The probability to free a table increase the longer they stay, with a maximum of 4 hours.